feat: poke list

+++ android/app/src/main/java/poke/rogue/helper/presentation/poketmon2/PokeMonItemClickListener.kt
@@ -1,5 +1,5 @@
 package poke.rogue.helper.presentation.poketmon2
 
 fun interface PokeMonItemClickListener {
-    fun onClickPokemon(pokemonId: Int)
+    fun onClickPokemon(pokemonId: Long)
 }
